server/form-pages/apply/risks-and-needs/health-needs/otherHealth.ts
import type { TaskListErrors } from '@approved-premises/ui'
import { Cas2Application as Application } from '@approved-premises/api'
import { Page } from '../../../utils/decorators'
import TaskListPage from '../../../taskListPage'

type OtherHealthBody = Record<string, never>

@Page({
  name: 'other-health',
  bodyProperties: [],
})
export default class OtherHealth implements TaskListPage {
  title = `Other health needs for ${this.application.person.name}`

  body: OtherHealthBody

  constructor(
    body: Partial<OtherHealthBody>,
    private readonly application: Application,
  ) {
    this.body = body as OtherHealthBody
  }

  previous() {
    return 'brain-injury'
  }

  next() {
    return ''
  }

  errors() {
    const errors: TaskListErrors<this> = {}

    return errors
  }

  response() {
    const response = {}

    return response
  }
}
